test: broken client test into multiple files
